+++ layer-api/src/main/java/org/layer/domain/actionItem/controller/ActionItemController.java
@@ -3,10 +3,10 @@ package org.layer.domain.actionItem.controller;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.layer.common.annotation.MemberId;
-import org.layer.domain.actionItem.controller.dto.ActionItemCreateRequest;
-import org.layer.domain.actionItem.controller.dto.MemberActionItemGetResponse;
-import org.layer.domain.actionItem.controller.dto.SpaceActionItemGetResponse;
-import org.layer.domain.actionItem.controller.dto.SpaceRetrospectActionItemGetResponse;
+import org.layer.domain.actionItem.controller.dto.request.ActionItemCreateRequest;
+import org.layer.domain.actionItem.controller.dto.response.MemberActionItemGetResponse;
+import org.layer.domain.actionItem.controller.dto.response.SpaceActionItemGetResponse;
+import org.layer.domain.actionItem.controller.dto.response.SpaceRetrospectActionItemGetResponse;
 import org.layer.domain.actionItem.service.ActionItemService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
