+++ layer-api/src/main/java/org/layer/domain/actionItem/controller/dto/response/MemberActionItemGetResponse.java
@@ -1,4 +1,4 @@
-package org.layer.domain.actionItem.controller.dto;
+package org.layer.domain.actionItem.controller.dto.response;
 
 import jakarta.validation.constraints.NotNull;
 import org.layer.domain.actionItem.dto.MemberActionItemResponse;
